# swagger.yaml
openapi: 3.0.0
info:
  title: Photo Cloud API Endpoints
  version: 1.0.0
  description: RESTful API Docs
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key needed to access endpoints
  schemas:
    ContentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the content
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the content was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the content was last updated
        name:
          type: string
          description: Name of the content
        description:
          type: string
          nullable: true
          description: Optional description of the content
        key:
          type: string
          description: Unique storage key for the content in the storage provider
        mimeType:
          type: string
          description: MIME type of the content (e.g., image/png, video/mp4)
        location:
          type: string
          description: URL or storage location of the content
        uploadCompletion:
          type: number
          description: Upload progress percentage (0 to 100)
        uploadError:
          type: string
          nullable: true
          description: Error message if the upload fails
        length:
          type: number
          nullable: true
          description: >-
            Length of the content (e.g., duration for videos, pages for
            documents)
        spaceId:
          type: string
          format: uuid
          description: The ID of the space to which this content belongs
        size:
          type: number
          description: Size of the content in bytes
      required:
        - id
        - createdAt
        - updatedAt
        - name
        - key
        - mimeType
        - location
        - uploadCompletion
        - spaceId
        - size
...
    UpdateContentRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the content. Maximum length of 255 characters.
        description:
          type: string
          description: A description of the content. Maximum length of 1000 characters.
        spaceId:
          type: integer
          description: The ID of the space associated with the content.
      required: []
...
security:
  - BearerAuth: []
    ApiKeyAuth: []
servers:
  - url: /api/v1
paths:
  /content/{content_id}:
    get:
      tags:
        - Content
      summary: Get content by ID
      description: Retrieve the content details by the specified content ID.
      parameters:
        - in: path
          name: content_id
          required: true
          schema:
            type: string
          description: The unique identifier of the content to retrieve.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved content details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '404':
          description: Content not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Content not found
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
    patch:
      tags:
        - Content
      summary: Update content by ID
      description: Update the details of a specific content by its content ID.
      parameters:
        - in: path
          name: content_id
          required: true
          schema:
            type: string
          description: The unique identifier of the content to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentRequest'
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully updated content details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '400':
          description: Invalid request body or missing parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid content data
        '404':
          description: Content not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Content not found
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
    delete:
      tags:
        - Content
      summary: Delete content by ID
      description: Delete the content with the specified content ID.
      parameters:
        - in: path
          name: content_id
          required: true
          schema:
            type: string
          description: The unique identifier of the content to delete.
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Successfully deleted content. No content returned.
        '404':
          description: Content not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Content not found
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
  /content/:
    get:
      tags:
        - Content
      summary: Find multiple contents with filters and pagination
      description: Retrieve a list of contents with pagination and filtering options.
      parameters:
        - in: query
          name: page_size
          required: true
          schema:
            type: integer
            minimum: 1
          description: The number of items per page.
        - in: query
          name: page_number
          required: true
          schema:
            type: integer
            minimum: 1
          description: The page number to retrieve.
        - in: query
          name: order
          required: true
          schema:
            type: string
            enum:
              - asc
              - desc
          description: The order in which to sort the results (ascending or descending).
        - in: query
          name: by
          required: true
          schema:
            type: string
          description: The field by which to sort the results.
        - in: query
          name: spaceId
          schema:
            type: string
          description: Filter content by space ID.
        - in: query
          name: mimeType
          schema:
            type: string
          description: Filter content by MIME type.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved a list of contents.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContentResponse'
                  pagination:
                    type: object
                    properties:
                      totalItems:
                        type: integer
                        example: 100
                      totalPages:
                        type: integer
                        example: 10
                      currentPage:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 10
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid query parameters
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
  /content/upload:
    post:
      tags:
        - Content
      summary: Upload multiple content files (images or videos)
      description: >-
        Upload multiple files (images or videos) to the specified space. The
        files should be sent via form-data with the key 'files'.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: The files to upload (multiple files).
                  required: true
                spaceId:
                  type: string
                  description: The ID of the space where the content will be uploaded.
                  required: true
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Successfully uploaded the content files with no content returned.
        '400':
          description: Bad request, invalid files or missing spaceId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid file types or missing spaceId
        '401':
          description: Unauthorized, invalid or missing authentication token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
...